Change in header de-duplication; Refactoring (#264)
* Change in header de-duplication

* enforce utf8 encoding

* more code refactoring
